refactor(frontend): extract default components
