+++ frontend/default/index.jsx
@@ -2,6 +2,17 @@
     import { THEMES, ALERT_CONFIG, TYPE_LABELS } from "./constants.js";
     import { REPORT_DATA } from "./data.js";
     import { ThemeContext } from "./theme.js";
+    import { filterMissingBins, hasValidMetrics } from "./helpers.js";
+    import { AlertBadge } from "./components/AlertBadge.jsx";
+    import { CategoricalChart } from "./components/CategoricalChart.jsx";
+    import { CorrelationMatrixComponent } from "./components/CorrelationMatrixComponent.jsx";
+    import { ExtremeValuesList } from "./components/ExtremeValuesList.jsx";
+    import { MetricRow } from "./components/MetricRow.jsx";
+    import { MetricsSection } from "./components/MetricsSection.jsx";
+    import { NullityMatrix } from "./components/NullityMatrix.jsx";
+    import { NumericHistogram } from "./components/NumericHistogram.jsx";
+    import { ScatterPlot } from "./components/ScatterPlot.jsx";
+    import { StatCard } from "./components/StatCard.jsx";
 
     const { useState, useMemo, useEffect, createContext, useContext } = React;
     const { 
@@ -11,471 +22,6 @@
       Fingerprint, Lock, Copy, FileWarning, TextSelect, X, Layers,
       Sun, Moon, Contrast
     } = LucideReact;
-
-    // Helper: Safely parse different matrix JSON structures
-    const parseMatrixData = (data) => {
-      if (!data) return { columns: [], matrix: [] };
-      
-      // 1) Handle { matrix: [ {col1: v1, ...}, ... ] } - new standard
-      if (data.matrix && Array.isArray(data.matrix) && data.matrix.length > 0 && typeof data.matrix[0] === 'object' && !Array.isArray(data.matrix[0])) {
-          const columns = Object.keys(data.matrix[0]);
-          const matrix = data.matrix.map(row => columns.map(col => row[col]));
-          return { columns, matrix };
-      }
-
-      // 2) If it's already an array, parse it as array-of-objects
-      if (Array.isArray(data)) {
-        if (data.length === 0) return { columns: [], matrix: [] };
-        const columns = Object.keys(data[0]);
-        const matrix = data.map(row => columns.map(col => row[col]));
-        return { columns, matrix };
-      }
-
-      // 3) Handle { columns, matrix } or { columns, values } standard shapes
-      if (data.columns) {
-        if (data.matrix) return { columns: data.columns, matrix: data.matrix };
-        if (data.values) return { columns: data.columns, matrix: data.values };
-      }
-      
-      return { columns: [], matrix: [] };
-    };
-
-    const ScatterPlot = ({ data, xLabel, yLabel }) => {
-      const { theme } = useContext(ThemeContext);
-      if (!data || data.length === 0) return <div className={`h-96 flex items-center justify-center ${theme.textSub} italic border ${theme.border} rounded-xl`}>No interaction data for this pair</div>;
-
-      const margin = { top: 20, right: 20, bottom: 40, left: 60 };
-      const width = 800;
-      const height = 400;
-
-      const xValues = data.map(d => d.x);
-      const yValues = data.map(d => d.y);
-      const xMin = Math.min(...xValues);
-      const xMax = Math.max(...xValues);
-      const yMin = Math.min(...yValues);
-      const yMax = Math.max(...yValues);
-
-      // Add 5% padding to scales
-      const xRange = xMax - xMin || 1;
-      const yRange = yMax - yMin || 1;
-      const xSMin = xMin - xRange * 0.05;
-      const xSMax = xMax + xRange * 0.05;
-      const ySMin = yMin - yRange * 0.05;
-      const ySMax = yMax + yRange * 0.05;
-
-      const xScale = (val) => margin.left + ((val - xSMin) / (xSMax - xSMin)) * (width - margin.left - margin.right);
-      const yScale = (val) => height - margin.bottom - ((val - ySMin) / (ySMax - ySMin)) * (height - margin.top - margin.bottom);
-
-      return (
-        <div className="relative">
-          <svg viewBox={`0 0 ${width} ${height}`} className={`w-full h-auto ${theme.bgCardAlt} rounded-xl border ${theme.border}`}>
-            {/* Grid lines */}
-            {[0, 0.25, 0.5, 0.75, 1].map(v => (
-              <React.Fragment key={v}>
-                <line 
-                  x1={margin.left} y1={yScale(ySMin + v * (ySMax - ySMin))} 
-                  x2={width - margin.right} y2={yScale(ySMin + v * (ySMax - ySMin))} 
-                  stroke={theme.border} strokeDasharray="4"
-                />
-                <line 
-                  x1={xScale(xSMin + v * (xSMax - xSMin))} y1={margin.top} 
-                  x2={xScale(xSMin + v * (xSMax - xSMin))} y2={height - margin.bottom} 
-                  stroke={theme.border} strokeDasharray="4"
-                />
-              </React.Fragment>
-            ))}
-
-            {/* Axes */}
-            <line x1={margin.left} y1={height - margin.bottom} x2={width - margin.right} y2={height - margin.bottom} stroke={theme.textMuted} strokeWidth="2" />
-            <line x1={margin.left} y1={margin.top} x2={margin.left} y2={height - margin.bottom} stroke={theme.textMuted} strokeWidth="2" />
-
-            {/* Points */}
-            {data.map((d, i) => (
-              <circle 
-                key={i} 
-                cx={xScale(d.x)} 
-                cy={yScale(d.y)} 
-                r="3" 
-                fill="rgba(59, 130, 246, 0.6)" 
-                className="hover:fill-blue-400 transition-colors"
-              >
-                <title>{`(${d.x.toFixed(4)}, ${d.y.toFixed(4)})`}</title>
-              </circle>
-            ))}
-
-            {/* Labels */}
-            <text x={width / 2} y={height - 5} textAnchor="middle" className={`text-xs font-bold fill-current ${theme.textMuted}`}>{xLabel}</text>
-            <text x={15} y={height / 2} textAnchor="middle" transform={`rotate(-90, 15, ${height / 2})`} className={`text-xs font-bold fill-current ${theme.textMuted}`}>{yLabel}</text>
-          </svg>
-        </div>
-      );
-    };
-
-    // Filter helper to remove None/Null/NaN bins from histogram charts
-    const filterMissingBins = (bins, counts) => {
-      if (!bins || !counts || !Array.isArray(bins) || !Array.isArray(counts)) {
-        return { filteredBins: bins || [], filteredCounts: counts || [] };
-      }
-      const missingLabels = ['none', 'null', 'nan', 'undefined', ''];
-      const filteredBins = [];
-      const filteredCounts = [];
-      for (let i = 0; i < bins.length; i++) {
-        const binVal = bins[i];
-        if (binVal === null || binVal === undefined) continue;
-        
-        let binStr = "";
-        try {
-          binStr = String(binVal).toLowerCase().trim();
-        } catch (e) {
-          binStr = "";
-        }
-        
-        // Don't filter out range labels like "[10.0, 20.0]"
-        if (!missingLabels.includes(binStr)) {
-          filteredBins.push(binVal);
-          filteredCounts.push(counts[i]);
-        }
-      }
-      return { filteredBins, filteredCounts };
-    };
-
-    const StatCard = ({ title, value, icon: Icon, sub, color="blue" }) => {
-      const { theme } = useContext(ThemeContext);
-      return (
-        <div className={`${theme.bgCard} border ${theme.border} p-5 rounded-2xl shadow-xl hover:shadow-${color}-900/10 transition-shadow`}>
-          <div className="flex justify-between items-start mb-3">
-            <span className={`${theme.textMuted} text-sm font-medium`}>{title}</span>
-            <div className={`p-2 rounded-lg bg-${color}-500/10 text-${color}-400`}>
-              <Icon size={18} />
-            </div>
-          </div>
-          <div className={`text-2xl font-mono font-bold ${theme.textPrimary}`}>{value}</div>
-          {sub && <div className={`text-xs ${theme.textSub} mt-1`}>{sub}</div>}
-        </div>
-      );
-    };
-
-    const NumericHistogram = ({ bins, counts, height = 60, showTooltip = true }) => {
-      const { theme } = useContext(ThemeContext);
-      if (!counts || counts.length === 0) return <div className={`h-full w-full ${theme.bgCardAlt} rounded-lg flex items-center justify-center text-xs ${theme.textSub}`}>No chart data</div>;
-      
-      // Ensure all counts are numbers and filter out any weirdness
-      const numericCounts = counts.map(c => Number(c) || 0);
-      const max = Math.max(...numericCounts, 1);
-      
-      return (
-        <div className={`flex items-end gap-0.5 w-full ${theme.bgCardAlt} p-2 rounded-lg`} style={{ height }}>
-          {numericCounts.map((c, i) => {
-            let binLabel = bins?.[i];
-            if (binLabel !== undefined && !String(binLabel).includes('[') && !Number.isNaN(Number(binLabel))) {
-              binLabel = Number(binLabel) % 1 !== 0 ? Number(binLabel).toFixed(2) : binLabel;
-            }
-
-            return (
-              <div 
-                key={i} 
-                className="bg-blue-500/60 hover:bg-blue-400 transition-all duration-200 rounded-t-sm flex-1 relative group"
-                style={{ height: `${(c / max) * 100}%` }}
-              >
-                {showTooltip && (
-                  <div className={`absolute bottom-full mb-2 left-1/2 -translate-x-1/2 ${theme.bgCard} border ${theme.border} ${theme.textPrimary} text-[10px] px-2 py-1 rounded hidden group-hover:block whitespace-nowrap z-10 pointer-events-none shadow-xl`}>
-                    <div className={`${theme.textMuted} mb-0.5`}>{binLabel ? `Bin: ${binLabel}` : 'Value'}</div>
-                    <div className="font-mono text-blue-300">{c.toLocaleString()} count</div>
-                  </div>
-                )}
-              </div>
-            );
-          })}
-        </div>
-      );
-    };
-
-    const CategoricalChart = ({ bins, counts, limit = 10 }) => {
-      const { theme } = useContext(ThemeContext);
-      if (!bins || !counts || counts.length === 0) return <div className={`p-4 text-center text-xs ${theme.textSub}`}>No data distribution data</div>;
-      const max = Math.max(...counts);
-      const displayLimit = limit;
-      const topBins = bins.slice(0, displayLimit);
-      const topCounts = counts.slice(0, displayLimit);
-
-      return (
-        <div className={`space-y-3 ${theme.bgCardAlt} p-4 rounded-xl border ${theme.borderAlt}`}>
-          {topBins.map((bin, i) => (
-            <div key={i} className="flex items-center gap-3 text-xs">
-              <div className={`w-24 truncate ${theme.textMuted} font-medium text-right`} title={String(bin)}>
-                {String(bin)}
-              </div>
-              <div className={`flex-1 h-5 ${theme.bgMain} rounded overflow-hidden relative group border ${theme.border}`}>
-                <div 
-                  className="h-full bg-blue-500/50 group-hover:bg-blue-400/80 transition-all rounded-r"
-                  style={{ width: `${(topCounts[i] / (max || 1)) * 100}%` }}
-                />
-              </div>
-              <div className={`w-12 font-mono ${theme.textSecondary} text-right`}>{topCounts[i].toLocaleString()}</div>
-            </div>
-          ))}
-          {bins.length > displayLimit && (
-            <div className={`text-center text-[10px] ${theme.textSub} mt-4 border-t ${theme.borderAlt} pt-2`}>
-              Showing top {displayLimit} of {bins.length} distinct categories
-            </div>
-          )}
-        </div>
-      );
-    };
-
-    const AlertBadge = ({ alert }) => {
-      const { theme } = useContext(ThemeContext);
-      const config = ALERT_CONFIG[alert.alert_type] || ALERT_CONFIG.DEFAULT;
-      const Icon = config.icon;
-      return (
-        <span className={`inline-flex items-center gap-1.5 px-2 py-0.5 rounded text-[10px] font-bold uppercase tracking-wider border ${config.color} ${config.bg} ${config.border}`} title={alert.message}>
-          <Icon size={12} />
-          {alert.alert_type}
-        </span>
-      );
-    };
-
-    const ExtremeValuesList = ({ title, values }) => {
-      const { theme } = useContext(ThemeContext);
-      if (!values || values.length === 0) return null;
-      return (
-        <div className="flex-1">
-          <h5 className={`text-[10px] font-bold ${theme.textSub} uppercase tracking-widest mb-3`}>{title}</h5>
-          <div className="space-y-1.5">
-            {values.map((v, i) => (
-              <div key={i} className={`flex items-center gap-2 text-xs ${theme.bgCardAlt} px-2 py-1.5 rounded border ${theme.border}`}>
-                <span className="text-slate-600 font-mono w-4">{i+1}.</span>
-                <span className={`${theme.textSecondary} font-mono truncate`} title={String(v)}>{String(v)}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-      );
-    };
-
-    const MetricRow = ({ label, val, isPercent = false }) => {
-      const { theme } = useContext(ThemeContext);
-      if (val === undefined || val === null || val === 'null' || val === 'N/A' || (typeof val === 'number' && Number.isNaN(val))) return null;
-      
-      let formattedVal = val;
-      if (isPercent && typeof val === 'number') {
-        formattedVal = `${(val * 100).toFixed(2)}%`;
-      } else if (typeof val === 'number') {
-        if (val !== 0 && Math.abs(val) < 0.001) {
-          formattedVal = val.toExponential(3);
-        } else {
-          formattedVal = val.toLocaleString(undefined, { maximumFractionDigits: 4 });
-        }
-      } else if (typeof val === 'boolean') {
-        formattedVal = val ? 'True' : 'False';
-      } else if (val === 'Skipped') {
-        formattedVal = 'Skipped';
-      }
-
-      return (
-        <div className={`flex justify-between items-center py-1.5 border-b ${theme.borderAlt} last:border-0 hover:bg-slate-800/20 transition-all duration-200 px-1 rounded`}>
-          <span className={`text-xs ${theme.textMuted}`}>{label}</span>
-          <span className={`text-xs font-mono font-bold ${theme.textSecondary}`}>
-            {String(formattedVal)}
-          </span>
-        </div>
-      );
-    };
-
-    const MetricsSection = ({ title, metrics }) => {
-      const { theme } = useContext(ThemeContext);
-      const validMetrics = metrics.filter(m => m.val !== undefined && m.val !== null && m.val !== 'null' && m.val !== 'N/A' && !(typeof m.val === 'number' && Number.isNaN(m.val)));
-      if (validMetrics.length === 0) return null;
-
-      return (
-        <div>
-          <h5 className={`text-[10px] font-bold ${theme.textMuted} uppercase tracking-widest mb-3`}>{title}</h5>
-          <div>
-            {validMetrics.map((m, i) => (
-              <MetricRow key={i} label={m.label} val={m.val} isPercent={m.isPercent} />
-            ))}
-          </div>
-        </div>
-      );
-    };
-
-    const hasValidMetrics = (metrics) => {
-      return metrics.some(m => m.val !== undefined && m.val !== null && m.val !== 'null' && m.val !== 'N/A' && !(typeof m.val === 'number' && Number.isNaN(m.val)));
-    };
-
-    // Sub-component for Nullity Matrix Rendering (Missingno style SVG)
-    const NullityMatrix = ({ data, metadata }) => {
-      const { theme } = useContext(ThemeContext);
-      // Data is now a list of dicts: [ {col1: true, col2: false}, ... ]
-      if (!data || !Array.isArray(data) || data.length === 0) return null;
-
-      const columns = Object.keys(data[0]);
-      const numRows = data.length;
-      const numCols = columns.length;
-      
-      // Convert list of dicts to row-major values array for processing
-      const values = data.map(row => columns.map(c => row[c]));
-
-      const svgWidth = 1000;
-      const svgHeight = 400;
-      const colWidth = svgWidth / numCols;
-      const rowHeight = svgHeight / numRows;
-
-      // Determine complete (non-missing) elements per row for the sparkline
-      const rowCompleteness = values.map(row => row.filter(isMissing => !isMissing).length);
-
-      // Helper to consolidate consecutive present data blocks for efficient SVG rendering
-      const getBlocks = (cIdx) => {
-          const blocks = [];
-          let inBlock = false;
-          let start = 0;
-          for (let r = 0; r < numRows; r++) {
-              // 'values' has true for missing. We draw dark blocks for PRESENT data.
-              const isPresent = !values[r][cIdx];
-              if (isPresent && !inBlock) {
-                  inBlock = true;
-                  start = r;
-              } else if (!isPresent && inBlock) {
-                  inBlock = false;
-                  blocks.push({ start, end: r - 1 });
-              }
-          }
-          if (inBlock) {
-              blocks.push({ start, end: numRows - 1 });
-          }
-          return blocks;
-      };
-
-      const sparklinePoints = rowCompleteness.map((count, rIdx) => 
-         `${(count / numCols) * 100},${rIdx * rowHeight + rowHeight / 2}`
-      ).join(' ');
-      const sparkPoly = `0,${rowHeight / 2} ${sparklinePoints} 0,${(numRows - 1) * rowHeight + rowHeight / 2}`;
-
-      return (
-        <div className={`mt-12 pt-8 border-t ${theme.border}`}>
-          <h3 className={`font-bold text-xl mb-2 ${theme.textPrimary}`}>Nullity Matrix</h3>
-          
-          {metadata?.truncated && (
-            <div className="mb-4 text-amber-400 text-sm font-medium flex items-center gap-2">
-              <AlertTriangle size={16} />
-              Displaying first {metadata.limit} columns out of {metadata.original_count} total.
-            </div>
-          )}
-
-          <p className={`text-sm ${theme.textSub} mb-6`}>Dense visualization of data completeness (first {numRows} rows). Gaps indicate missing data.</p>
-
-          <div className="flex gap-4 items-end" style={{ minWidth: `${Math.max(1000, numCols * 25)}px` }}>
-            <div className="flex-1 relative">
-              {/* Column Labels */}
-              <div className="flex mb-2">
-                 {columns.map((c, i) => (
-                    <div key={c} style={{ width: `${100/numCols}%` }} className="flex justify-start h-32 relative overflow-visible">
-                       <span className={`absolute bottom-1 left-1/2 origin-bottom-left text-[10px] font-mono ${theme.textMuted} whitespace-nowrap ${numCols > 15 ? '-rotate-90' : '-rotate-45'}`}>{c}</span>
-                    </div>
-                 ))}
-              </div>
-              
-              {/* SVG Matrix */}
-              <svg viewBox={`0 0 ${svgWidth} ${svgHeight}`} preserveAspectRatio="none" className={`w-full h-[500px] ${theme.bgInput} border ${theme.border} shadow-inner rounded-sm`}>
-                {columns.map((_, cIdx) => (
-                   <g key={cIdx} transform={`translate(${cIdx * colWidth}, 0)`}>
-                      {getBlocks(cIdx).map((block, i) => (
-                         <rect
-                            key={i}
-                            x={colWidth * 0.05}
-                            y={block.start * rowHeight}
-                            width={colWidth * 0.9}
-                            height={(block.end - block.start + 1) * rowHeight}
-                            fill={theme.id === 'light' ? '#475569' : '#cbd5e1'}
-                            className="transition-all hover:fill-blue-400"
-                         />
-                      ))}
-                   </g>
-                ))}
-              </svg>
-            </div>
-
-            {/* Completeness Sparkline */}
-            <div className="w-20 h-[500px] relative flex flex-col items-center">
-              <div className={`text-[10px] ${theme.textSub} font-mono text-center absolute -top-8 w-full`}>Data<br/>Completeness</div>
-              <svg viewBox={`0 0 100 ${svgHeight}`} preserveAspectRatio="none" className={`w-full h-full border-r border-t border-b ${theme.borderAlt} ${theme.bgCardAlt} rounded-r-sm`}>
-                <polygon points={sparkPoly} fill="rgba(59, 130, 246, 0.15)" />
-                <polyline points={sparklinePoints} fill="none" stroke="#3b82f6" strokeWidth="1.5" />
-              </svg>
-              <div className={`absolute top-2 right-2 text-xs font-mono ${theme.textMuted}`}>{numCols}</div>
-              <div className={`absolute bottom-2 right-2 text-xs font-mono ${theme.textMuted}`}>0</div>
-            </div>
-          </div>
-        </div>
-      );
-    }
-
-    // Sub-component for rendering standardized Correlation Matrices
-    const CorrelationMatrixComponent = ({ data }) => {
-      const { theme } = useContext(ThemeContext);
-      const { columns, matrix } = useMemo(() => parseMatrixData(data), [data]);
-      
-      if (!columns || columns.length === 0 || !matrix || matrix.length === 0) {
-        return <div className={`py-12 text-center ${theme.textSub}`}>No data available for this metric.</div>;
-      }
-
-      return (
-        <div className="overflow-x-auto pb-8">
-          <table className="min-w-max border-collapse">
-            <thead>
-              <tr>
-                <th className="p-1"></th>
-                {columns.map(c => (
-                  <th key={c} className={`p-2 font-mono font-bold ${theme.textMuted} h-32 align-bottom`}>
-                    <div className={`transform origin-bottom-left ml-8 truncate w-32 ${columns.length > 15 ? '-rotate-90' : '-rotate-45'}`}>
-                      {c}
-                    </div>
-                  </th>
-                ))}
-              </tr>
-            </thead>
-            <tbody>
-              {matrix.map((row, i) => (
-                <tr key={i}>
-                  <td className={`p-2 font-mono font-bold ${theme.textMuted} text-right pr-8 whitespace-nowrap min-w-[150px]`}>
-                    {columns[i]}
-                  </td>
-                  {row.map((cell, j) => {
-                    const isMissing = cell === null || cell === undefined || Number.isNaN(cell);
-                    const bgOpacity = isMissing ? 0 : Math.min(Math.abs(cell), 1);
-                    // Blue for positive correlation, Red for negative
-                    const bgColor = isMissing ? 'transparent' : cell >= 0 ? `rgba(59, 130, 246, ${bgOpacity})` : `rgba(239, 68, 68, ${bgOpacity})`;
-                    
-                    return (
-                      <td 
-                        key={j} 
-                        className={`w-16 h-16 border ${theme.border} text-center hover:scale-105 transition-transform cursor-crosshair relative group`} 
-                        style={{ backgroundColor: bgColor }}
-                      >
-                        {isMissing ? (
-                          <span className={`${theme.textMuted} text-[10px] italic`}>NaN</span>
-                        ) : (
-                          <>
-                            <span className={`font-mono text-xs font-bold ${bgOpacity > 0.5 ? 'text-white' : (theme.id === 'light' ? 'text-slate-900' : 'text-slate-400')}`}>
-                              {cell === 1 ? '1.0' : cell.toFixed(2)}
-                            </span>
-                            <div className={`absolute z-10 hidden group-hover:block bottom-full left-1/2 -translate-x-1/2 mb-2 px-2 py-1 ${theme.bgCard} ${theme.textPrimary} text-xs rounded whitespace-nowrap shadow-xl border ${theme.border}`}>
-                              {columns[i]} × {columns[j]}
-                              <br/>
-                              <span className="font-mono text-blue-300">{cell.toFixed(4)}</span>
-                            </div>
-                          </>
-                        )}
-                      </td>
-                    );
-                  })}
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      );
-    };
 
     function AppContent() {
       const { theme, setTheme } = useContext(ThemeContext);
